% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\docType{methods}
\name{getAssociationWithTSS}
\alias{getAssociationWithTSS}
\alias{getAssociationWithTSS,annotationByGenicParts-method}
\title{Get distance to nearest TSS and gene id from annotationByGenicParts}
\usage{
getAssociationWithTSS(x)

\S4method{getAssociationWithTSS}{annotationByGenicParts}(x)
}
\arguments{
\item{x}{an \code{\link[methylKit]{annotationByGenicParts}}  object}
}
\value{
a \code{\link{data.frame}} containing row number of the target 
features,distance of target to nearest TSS, TSS/Gene name, TSS strand
}
\description{
This accessor function gets the nearest TSS, its distance to target 
feature,strand and name of TSS/gene from annotationByGenicParts object
}
\examples{
data(methylKit)
gene.obj=readTranscriptFeatures(system.file("extdata", 
"refseq.hg18.bed.txt", package = "methylKit"))
ann=annotateWithGenicParts(methylDiff.obj, gene.obj)
df=getAssociationWithTSS(ann)
head(df)

}

