% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{applyTbxByChr}
\alias{applyTbxByChr}
\title{Apply a function on tabix files chromosome by chromosome}
\usage{
applyTbxByChr(tbxFile, chrs, dir, filename, return.type = c("tabix",
  "data.frame", "data.table"), FUN, ..., mc.cores = 1)
}
\arguments{
\item{tbxFile}{tabix file to read. a TabixFile object}

\item{chrs}{chromosome names. Based on chromosome names the chunks of 
tabix file
       will be read into the memory. If missing use all chromosome names 
       in tabix file.}

\item{dir}{directory to create temporary files and the resulting tabix file}

\item{filename}{the filename for the resulting tabix file, this should not be 
a path, just a file name.}

\item{return.type}{indicates the return type for the function}

\item{FUN}{function to apply to chrs, it takes a data.frame and returns a 
data.frame. First argument of the function should be a data frame}

\item{...}{parameters to be passed to FUN.}

\item{mc.cores}{number of cores to use in parallel (works only on UNIX based OS)}
}
\value{
either a path to a tabix or text file, or a data frame or data.table
}
\description{
The function reads a tabix file chromosome by chromosome and applies 
a function on each. 
The function (FUN argument) should apply on data.frames and return a 
data frame
as a result. The function is parallel applied to each chromosome 
and related arguments could be passed on to the function via ... argument.
}

