% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{applyTbxByOverlap}
\alias{applyTbxByOverlap}
\title{Serially apply a function on regions of tabix files}
\usage{
applyTbxByOverlap(tbxFile, ranges, chunk.size = 1e+06, dir, filename,
  return.type = c("tabix", "data.frame", "data.table"), FUN, ...)
}
\arguments{
\item{tbxFile}{tabix file to read. a TabixFile object}

\item{ranges}{a GRanges object specifying the regions}

\item{chunk.size}{number of rows to be taken as a chunk, default: 1e6}

\item{dir}{directory to create temporary files and the resulting tabix file}

\item{filename}{the filename for the resulting tabix file, this should not be 
a path, just a file name.}

\item{return.type}{indicates the return type for the function}

\item{FUN}{function to apply to chunks, it takes a data.frame and returns a 
data.frame. First argument of the function should be a data frame}

\item{...}{parameters to be passed to FUN.}
}
\value{
either a path to a tabix or text file, or a data frame or data.table
}
\description{
The function reads regions of a tabix file and applies a function on them. 
The function (FUN argument) should apply on data.frames and return a data frame
as a result. The function is serially applied to chunks (means no parallelization). 
However, the function FUN itself can be a parallelized function
and related arguments could be passed on to the function via ... argument.
}

