% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffMethDSS.R
\docType{methods}
\name{calculateDiffMethDSS}
\alias{calculateDiffMethDSS}
\title{calculate Differential Methylation with DSS}
\usage{
calculateDiffMethDSS(meth, slim=FALSE, num.cores=1)
}
\arguments{
\item{meth}{a methylBase object}

\item{slim}{default: FALSE. If set to TRUE, pvalues will be corrected with the methylKit-SLIM-function.}

\item{num.cores}{the number processors to be used}
}
\value{
a methylDiff object
}
\description{
This function provides an interface to the DSS package by Hao Wo at Emory University.
}
\section{Details}{

test
}
\examples{
data(methylKit)

dssDiffay <- calculateDiffMethDSS(methylBase.obj, slim=FALSE, num.cores=1)
}

