% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{catsub2tabix}
\alias{catsub2tabix}
\title{cat subfiles and make tabix}
\usage{
catsub2tabix(dir,pattern,filename,sort=F)
}
\arguments{
\item{dir}{directory of subfiles to be merged}

\item{pattern}{a pattern in the file names}

\item{filename}{output file name}

\item{sort}{sort list of subfiles in alpha, numerical way}
}
\description{
cat subfiles and make tabix
}

