% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{df2tabix}
\alias{df2tabix}
\title{make a tabix file from a data.frame or data.table part of a 
methylKit object}
\usage{
df2tabix(df,outfile)
}
\arguments{
\item{df}{data part of a methylKit object, either data.frame or data.table}

\item{outfile}{name of the output file}
}
\description{
make a tabix file from a data.frame or data.table part of a 
methylKit object
}

