% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{distance2nearestFeature}
\alias{distance2nearestFeature}
\title{distance2nearestFeature}
\usage{
distance2nearestFeature(g.idh, tss, full.output = FALSE)
}
\arguments{
\item{g.idh}{}

\item{tss}{}

\item{full.output}{}
}
\description{
distance2nearestFeature
}
\examples{
distance2nearestFeature(g.idh = g.idh, tss = tss, full.output = FALSE)
}

