% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{extract}
\alias{[,methylBase,ANY-method}
\alias{[,methylBase-method}
\alias{[,methylDiff,ANY-method}
\alias{[,methylDiff-method}
\alias{[,methylRaw-method}
\alias{extract}
\title{extract parts of methylRaw,methylBase and methylDiff data}
\usage{
\S4method{[}{methylRaw,ANY}(x, i, j)

\S4method{[}{methylBase,ANY}(x, i, j, drop)

\S4method{[}{methylDiff,ANY}(x, i, j)
}
\arguments{
\item{x}{an \code{\link{methylBase}},\code{\link{methylRaw}} or
\code{\link{methylDiff}} object}

\item{i}{a numeric or logical vector. This vector corresponds to bases or
regions contained in \code{methylKit} objects.The vector is used to
subset the data.}

\item{j}{This argument can not be used for the extraction of columns.
         As unintentional extraction of the columns will cause an error in
         the downstream analysis. Using this argument will cause an error.
          Use \code{\link[methylKit]{getData}} to access the data part of
          the objects.}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }
}
\description{
The function extracts part of the data and returns a new object.
}
\examples{
data(methylKit)

# selects first hundred rows, returns a methylRaw object
subset1=methylRawList.obj[[1]][1:100]

# selects first hundred rows, returns a methylBase object
subset2=methylBase.obj[1:100,]

# selects first hundred rows, returns a methylDiff object
subset3=methylDiff.obj[1:100,]

# This will get chromomsomes, will return a factor
# That means the resulting object will ceases to be a methylKit object
chrs=methylDiff.obj[[2]]
}

