\docType{methods}
\name{getAssembly}
\alias{"getAssembly,-methods"}
\alias{"getAssembly,ANY-method"}
\alias{"getAssembly,methylBase-method"}
\alias{getAssembly}
\title{get assembly of the genome}
\arguments{
  \item{x}{a methylBase object}
}
\value{
  the assembly string for the object
}
\description{
  get assembly of the genome
}

