\docType{methods}
\name{getContext}
\alias{getContext}
\alias{getContext,methylBase-method}
\alias{getContext,methylDiff-method}
\alias{getContext,methylRaw-method}
\title{get the context of methylation}
\arguments{
  \item{x}{a methylBase/methylRaw/methylDiff object}
}
\value{
  the context of methylation string
}
\description{
  get the context of methylation
}

