\docType{methods}
\name{getCorrelation}
\alias{"getCorrelation,-methods"}
\alias{"getCorrelation,methylBase-method"}
\alias{getCorrelation}
\title{get correlation between samples in methylBase object}
\arguments{
  \item{.Object}{a methylBase object}

  \item{plot}{scatterPlot if TRUE (deafult:False)}
}
\value{
  a correlation matrix object and plot scatterPlot
}
\description{
  get correlation between samples in methylBase object
}

