% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\docType{methods}
\name{getMembers}
\alias{getMembers}
\alias{getMembers,annotationByFeature-method}
\title{Get the membership slot of annotationByFeature}
\usage{
getMembers(x)

\S4method{getMembers}{annotationByFeature}(x)
}
\arguments{
\item{x}{an \code{\link[methylKit]{annotationByFeature}}  object}
}
\value{
a matrix showing overlap of target features with annotation features. 
1 for overlap, 0 for non-overlap. 
Each row in the matrix corresponds to a genomic feature that is annoted by 
one of the following functions:
\code{\link[methylKit]{annotateWithFeature}},
\code{\link[methylKit]{annotateWithFeatureFlank}},
\code{\link[methylKit]{annotateWithGenicParts}}
}
\description{
Membership slot defines the overlap of target features with annotation 
features as a matrix.
}
\examples{
data(methylKit)
cpg.obj=readFeatureFlank(system.file("extdata", "cpgi.hg18.bed.txt", 
package = "methylKit"),feature.flank.name=c("CpGi","shores"))

ann=annotateWithFeatureFlank(methylDiff.obj,cpg.obj$CpGi,cpg.obj$shores,
feature.name="CpGi",flank.name="Shores")
mat=getMembers(ann)
head(mat)

}

