% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{getTabixByChr}
\alias{getTabixByChr}
\title{get data from meth tabix for a given chr}
\usage{
getTabixByChr(tbxFile, chr = "chr10", return.type = c("data.table",
  "data.frame", "GRanges"))
}
\description{
get data from meth tabix for a given chr
}

