% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{getTabixByChunk}
\alias{getTabixByChunk}
\title{get data from already opened tabixfile for a given chunkSize}
\usage{
getTabixByChunk(tbxFile, chunk.size = 1e+06, return.type = c("data.table",
  "data.frame", "GRanges"))
}
\description{
get data from already opened tabixfile for a given chunkSize
}

