% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{getTabixByOverlap}
\alias{getTabixByOverlap}
\title{get data from meth tabix for a given set of regions}
\usage{
getTabixByOverlap(tbxFile, granges, return.type = "data.table")
}
\description{
get data from meth tabix for a given set of regions
}

