% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{headTabix}
\alias{headTabix}
\title{get data from meth tabix for a given number of rows}
\usage{
headTabix(tbxFile, nrow = 10, return.type = "data.table")
}
\description{
get data from meth tabix for a given number of rows
}

