% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methSeg.R
\name{joinSegmentNeighbours}
\alias{joinSegmentNeighbours}
\title{Join directly neighbouring segments produced by methSeg}
\usage{
joinSegmentNeighbours(res)
}
\arguments{
\item{res}{A \code{\link[GenomicRanges]{GRanges}} object with segment 
classification and information prudoced by \code{\link{methSeg}}}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with segment 
        classification and information.
}
\description{
Segmentation and clustering are two distinct steps in methSeg(), 
leading to adjacent segments of the same class. 
This leads to a bias segment length distributions, 
which is removed by joining those neighbours.
}
\seealso{
\code{\link{methSeg}}
}
