% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{makeMethTabix}
\alias{makeMethTabix}
\title{make tabix from a flat file where first 3 columns
are chr,start,end in that order}
\usage{
makeMethTabix(filepath,skip=0)
}
\arguments{
\item{filepath}{path to the uncompressed file}

\item{skip}{number of lines to skip}
}
\description{
make tabix from a flat file where first 3 columns
are chr,start,end in that order
}

