% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{mergeTabix}
\alias{mergeTabix}
\title{merge tabix files by chr, start,end, strand}
\usage{
mergeTabix(tabixList,dir,filename,mc.cores=1 ,all=FALSE)
}
\arguments{
\item{tabixList}{list of tabix files}

\item{dir}{working directory}

\item{filename}{the output file name}

\item{mc.cores}{number of multiple cores. If mc.cores>1 temporary files for
each chromsome will be created prior to cat, zipping and indexing the
single output file}

\item{all}{logical parameter passed to \code{\link{merge}} function

mergeTabix(tabixList,dir="~",filename="dump.meth.txt",mc.cores=1)}
}
\description{
merge tabix files by chr, start,end, strand
}

