% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.functions.R
\name{mergeTbxByChr}
\alias{mergeTbxByChr}
\title{merge the data tables for a given chr}
\usage{
mergeTbxByChr(chr, tabixList, dir, filename, parallel = FALSE, all = FALSE)
}
\arguments{
\item{chr}{chromosome to merge}

\item{tabixList}{list of tabix files}

\item{dir}{working directory}

\item{filename}{the output file name}

\item{parallel}{logical to determine internal processing of output}

\item{all}{logical parameter passed to \code{\link{merge}} function}
}
\description{
merge the data tables for a given chr
}

