% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methSeg.R
\docType{methods}
\name{methSeg}
\alias{methSeg}
\title{segment the methylation profile}
\usage{
methSeg(pathToFile, diagnostic.plot=TRUE, ...)
}
\arguments{
\item{obj}{\code{\link[GenomicRanges]{GRanges}}, \code{\link[methylKit]{methylRaw}} or \code{\link[methylKit]{methylDiff}} object to be segmented}

\item{diagnostic.plot}{if TRUE a diagnostic plot is plotted. The plot shows
metthylation and length statistics per segment group. In addition, it
shows diagnostics from mixture modeling: the density function estimated
and BIC criterion used to decide the optimum number of components
in mixture modeling.}

\item{...}{arguments to fastseg() function in fastseg package, or to densityFind
       in Mclust package, could be used to fine tune the segmentation algorithm.
       E.g. Increasing "alpha" will give more segments.
       Increasing "cyberWeight" will give also more segments.
       For more details see fastseg and Mclust documentation.}

\item{score.gr}{\code{\link[GenomicRanges]{GRanges}} objects containing methylation scores in meta-columns}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with segment classification and information.
       'seg.mean' column shows the mean methylation per segment.
       'seg.group' column shows the segment groups obtained by mixture modeling
}
\description{
The function uses fastseg() function to segment the profiles and
gaussian mixture modelling to cluster the segments into k components. Each
component ideally indicates quantitative classification of segments. Such
as high or low methylated regions.
}
\section{Details}{

       To be sure that the algorithm will work on your data, the object should have at least 5.000 rows
}

