% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methSeg.R
\docType{methods}
\name{methSeg2bed}
\alias{methSeg2bed}
\title{Export segments to BED files}
\usage{
methSeg2bed(segments,
  trackLine = "track name='meth segments' description='meth segments' itemRgb=On",
  filename = "data/H1.chr21.chr22.trial.seg.bed")
}
\arguments{
\item{segments}{\code{\link[GenomicRanges]{GRanges}} object with segment classification and information}

\item{trackLine}{UCSC browser trackline}

\item{filename}{name of the output data}
}
\value{
A BED files with the segmented data
which can be visualized in the UCSC browser
}
\description{
Export segments to BED files
}

