% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{methylBase.obj}
\alias{methylBase.obj}
\title{Example methylBase object.}
\format{methylBase.obj object stores the location and methylation 
information for bases that
  are covered in all samples. \code{methylBase} partially extends 
  \code{data.frame} S3 class.}
\description{
\code{methylBase}, \code{methylDiff} and \code{methylRawList}. 
 You can load the data using \code{data(methylKit)}
}

