% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{methylBaseDB.obj}
\alias{methylBaseDB.obj}
\title{Example methylBaseDB object.}
\format{methylBaseDB.obj object stores the location and methylation information for bases that
  are covered in all samples. \code{methylBaseDB} has the same functionality as \code{methylBase} class.}
\description{
methylKit has several objects. This data set includes examples of the following objects:
 \code{methylBase}, \code{methylDiff} and \code{methylRawList},
 aswell as their database counterparts \code{methylBaseDB}, \code{methylDiffDB} and \code{methylRawListDB}.
 You can load the data using \code{data(methylKit)}
}

