% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\docType{methods}
\name{methylKit-defunct}
\alias{methylKit-defunct}
\alias{annotate.WithFeature}
\alias{annotate.WithFeature.Flank}
\alias{annotate.WithGenicParts}
\alias{read.bed}
\alias{read.feature.flank}
\alias{read.transcript.features}
\alias{getFeatsWithTargetsStats}
\alias{getFlanks}
\alias{getMembers}
\alias{getTargetAnnotationStats}
\alias{plotTargetAnnotation}
\alias{read}
\alias{read.bismark}
\alias{adjust.methylC}
\alias{get.methylDiff}
\title{Deprecated/Defunct functions}
\usage{
annotate.WithFeature()

annotate.WithFeature.Flank()

annotate.WithGenicParts()

read.bed()

read.feature.flank()

read.transcript.features()

getFeatsWithTargetsStats()

getFlanks()

getMembers()

getTargetAnnotationStats()

plotTargetAnnotation()

read()

read.bismark()

adjust.methylC()

get.methylDiff()
}
\description{
These are deprecated or defunct functions. Most of them 
are replaced by genomation functions. See the vignette for 
examples on how to use genomation functions for annotation
purposes.
}
