\name{methylRawList-class}
\alias{methylRawList-class}
\title{An S4 class for holding a list of methylRaw objects.}
\description{
  This object stores the list of raw mehylation data that
  is read in through read function and extends data.frame
}
\section{Slots}{
  \describe{ \item{\code{treatment}:}{numeric vector
  denoting control and test samples} }
}

