\docType{data}
\name{methylRawList.obj}
\alias{methylRawList.obj}
\title{Example methylRawList object.}
\format{Methylation data from multiple the samples regardless
  of common coverage are stored in  methylRawList.obj object. \code{methylRawList} extends \code{list} S3 class}
\description{
  methylKit has several objects. This data set includes
  examples of the following objects: \code{methylBase},
  \code{methylDiff} and \code{methylRawList}. You can load
  the data using \code{data(methylKit)}
}
\keyword{datasets}

