% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylDBClasses.R
\docType{class}
\name{methylRawListDB-class}
\alias{methylRawListDB}
\alias{methylRawListDB-class}
\title{An S4 class for holding a list of methylRawDB objects.}
\description{
This class stores the list of  \code{\link[methylKit]{methylRawDB}} objects.
Functions such as \code{lapply} can be used on this list. It extends
 \code{\link[base]{list}} class. This object is primarily produced
by \code{\link[methylKit]{read}} function.
}
\section{Slots}{
\describe{
                 \item{\code{treatment}}{numeric vector denoting control and test samples}
                 \item{\code{.Data}}{a list of \code{\link{methylRawDB}} objects  } 
               }
}
\examples{
file.list=list( system.file("extdata", "test1.myCpG.txt", package = "methylKit"),
                system.file("extdata", "test2.myCpG.txt", package = "methylKit"),
                system.file("extdata", "control1.myCpG.txt", package = "methylKit"),
                system.file("extdata", "control2.myCpG.txt", package = "methylKit") )

methylRawListDB.obj <- read(file.list,sample.id = list("test1","test2","ctrl1","ctrl2"),
                            assembly = "hg18",treatment = c(1,1,0,0),
                            dbtype = "tabix",dbdir = "methylDB")

#applying functions designed for methylRawDB on methylRawListDB object
lapply(methylRawListDB.obj,"getAssembly")


# remove Database again
rm(methylRawListDB.obj)
unlink("methylDB",recursive=TRUE)

}

