% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{methylRawListDB.obj}
\alias{methylRawListDB.obj}
\title{Example methylRawListDB object.}
\format{Methylation data from multiple the samples regardless
  of common coverage are stored in  methylRawListDB.obj object as flat file database.
  \code{methylRawListDB} extends \code{list} S3 class.}
\description{
methylKit has several objects. This data set includes examples of the following objects:
 \code{methylBase}, \code{methylDiff} and \code{methylRawList},
 aswell as their database counterparts \code{methylBaseDB}, \code{methylDiffDB} and \code{methylRawListDB}.
 You can load the data using \code{data(methylKit)}
}

