% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylDBClasses.R
\docType{methods}
\name{readMethylDB}
\alias{readMethylDB}
\title{coerce methylKit objects from memory to flat file database objects}
\usage{
readMethylDB(dbpath)
}
\arguments{
\item{obj}{an \code{\link{methylBase}},
\code{\link{methylRaw}},
\code{\link{methylRawList}} or an \code{\link{methylDiff}} object}

\item{dbdir}{directory where flat file database(s) should be stored, 
defaults to getwd(), working directory.}
}
\value{
an \code{\link{methylBaseDB}},\code{\link{methylRawDB}},
\code{\link{methylRawListDB}} or an \code{\link{methylDiffDB}} object
}
\description{
The function converts in-memory methylKit objects to methylDB objects
}
\examples{
\dontrun{
data(methylKit)


makeMethylDB(methylBase.obj,"my/path")
}

}
