% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methylDBClasses.R
\docType{methods}
\name{selectByOverlap}
\alias{selectByOverlap}
\alias{selectByOverlap,methylBaseDB-method}
\alias{selectByOverlap,methylDiffDB-method}
\alias{selectByOverlap,methylRawDB-method}
\alias{selectByOverlap,methylRawListDB-method}
\title{selects records of methylDB objects lying inside a GRanges range}
\usage{
selectByOverlap(object,ranges)

\S4method{selectByOverlap}{methylRawDB}(object, ranges)

\S4method{selectByOverlap}{methylRawListDB}(object, ranges)

\S4method{selectByOverlap}{methylBaseDB}(object, ranges)

\S4method{selectByOverlap}{methylDiffDB}(object, ranges)
}
\arguments{
\item{object}{an \code{\link{methylBaseDB}},\code{\link{methylRawDB}} or \code{\link{methylDiffDB}} object}

\item{ranges}{a GRanges object specifying the regions of interest}
}
\value{
a \code{\link{methylBase}},\code{\link{methylRaw}} or
          \code{\link{methylDiff}} object depending on the input object.
}
\description{
The function selects records from a \code{\link{methylBaseDB}}, \code{\link{methylRawDB}} or \code{\link{methylDiffDB}} object
that lie inside the regions given by \code{ranges} of class \code{GRanges} and returns an object of class
\code{\link{methylBase}}, \code{\link{methylRaw}} or \code{\link{methylDiff}}
}
\examples{
data(methylKit)

# define the windows of interest as a GRanges object, this can be any set
# of genomic locations
library(GenomicRanges)
my.win=GRanges(seqnames="chr21",
ranges=IRanges(start=seq(from=9764513,by=10000,length.out=20),width=5000) )

# selects the records that lie inside the regions
myRaw <- selectByOverlap(methylRawListDB.obj[[1]],my.win)

# selects the records that lie inside the regions
myBase <- selectByOverlap(methylBaseDB.obj,my.win)

# selects the records that lie inside the regions
myDiff <- selectByOverlap(methylDiffDB.obj,my.win)
}

