% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMethObject.R
\docType{methods}
\name{updateMethObject,methylBase-method}
\alias{updateMethObject,methylBase-method}
\title{@aliases updateMethObject,methylBase-method
 @rdname updateMethObject}
\usage{
\S4method{updateMethObject}{methylBase}(object)
}
\description{
@aliases updateMethObject,methylBase-method
 @rdname updateMethObject
}

