% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMethObject.R
\docType{methods}
\name{updateMethObject,methylDiff-method}
\alias{updateMethObject,methylDiff-method}
\title{@aliases updateMethObject,methylDiff-method
 @rdname updateMethObject}
\usage{
\S4method{updateMethObject}{methylDiff}(object)
}
\description{
@aliases updateMethObject,methylDiff-method
 @rdname updateMethObject
}

