% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMethObject.R
\docType{methods}
\name{updateMethObject,methylRaw-method}
\alias{updateMethObject,methylRaw-method}
\title{@aliases updateMethObject,methylRaw-method
 @rdname updateMethObject}
\usage{
\S4method{updateMethObject}{methylRaw}(object)
}
\description{
@aliases updateMethObject,methylRaw-method
 @rdname updateMethObject
}

