% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMethObject.R
\name{updateMethObject}
\alias{updateMethObject}
\title{update methylKit objects 
 
The method updates object from earlier versions (<v0.9.1) to latest object.}
\usage{
updateMethObject(object)
}
\arguments{
\item{object}{a methylKit object: methylRaw, methylRawList, methylBase or methylDiff}
}
\value{
\code{\link{methylRaw}},\code{\link{methylDiff}},\code{\link{methylBase}}
or \code{\link{methylRawList}} object


 @export 
 @docType methods
 @rdname updateMethObject
}
\description{
update methylKit objects 
 
The method updates object from earlier versions (<v0.9.1) to latest object.
}

