% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\docType{class}
\name{annotationByFeature-class}
\alias{annotationByFeature}
\alias{annotationByFeature-class}
\title{An S4 class for  overlap of target features with a generic annotation}
\description{
This object is desgined to hold statistics and information about genomic feature overlaps it extends \code{\link{list}} class.
}
\section{Slots}{
\describe{
                 \item{\code{members}}{a matrix showing overlap of target features with annotation genomic features}
                 \item{\code{annotation}}{a named vector of percentages of overlap between feature and annotation}'
                 \item{\code{precedence}}{a named vector of percentages of overlap between feature and annotation}
                 \item{\code{num.annotation}}{a named vector of numbers of overlap between feature and annotation}
                 \item{\code{num.precedence}}{a named vector of numbers of overlap between feature and annotation}
                 \item{\code{no.of.OlapFeat}}{vector}
                 \item{\code{perc.of.OlapFeat}}{vector}
}
}
\examples{
data(methylKit)
cpg.obj=read.feature.flank(system.file("extdata", "cpgi.hg18.bed.txt", package = "methylKit"),feature.flank.name=c("CpGi","shores"))

# the following function returns annotationByFeature object
ann=annotate.WithFeature.Flank(methylDiff.obj,cpg.obj$CpGi,cpg.obj$shores,feature.name="CpGi",flank.name="Shores")
ann

}
\seealso{
see \code{\link[methylKit]{annotate.WithFeature.Flank}} and \code{\link[methylKit]{annotate.WithFeature}} on how to create this object.
         see following functions that operates on this object:
           \code{\link[methylKit]{getMembers}}, \code{\link[methylKit]{getTargetAnnotationStats}},
           \code{\link[methylKit]{getFeatsWithTargetsStats}}, \code{\link[methylKit]{plotTargetAnnotation}}
}

