% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffMeth.R
\docType{methods}
\name{calculateDiffMeth}
\alias{calculateDiffMeth}
\alias{calculateDiffMeth,methylBase-method}
\title{Calculate differential methylation statistics}
\usage{
calculateDiffMeth(.Object,covariates,overdispersion=c("none","MN","shrinkMN"),
adjust=c("SLIM","holm","hochberg","hommel","bonferroni","BH","BY","fdr","none","qvalue"),
effect=c("wmean","mean","predicted"),parShrinkNM=list(),
test=c("F","Chisq"),mc.cores=1)



#covariates=NULL,overdispersion=c("none"),adjust=c("SLIM"),
#                          effect=c("wmean"),parShrinkNM=list(),test=c("F"),mc.cores=1)

\S4method{calculateDiffMeth}{methylBase}(.Object, covariates = NULL,
  overdispersion = c("none", "MN", "shrinkMN"), adjust = c("SLIM", "holm",
  "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none", "qvalue"),
  effect = c("wmean", "mean", "predicted"), parShrinkNM = list(),
  test = c("F", "Chisq"), mc.cores = 1)
}
\arguments{
\item{.Object}{a methylBase object to calculate differential methylation}

\item{covariates}{a data.frame containing covariates, which should be included in the test.}

\item{overdispersion}{If set to "none"(default), no overdispersion correction will be attempted.
If set to "MN", basic overdispersion correction will be applied and
if set to "shrinkMN", overdisperison correction with squeezeVar()
from the limma-package will be applied.}

\item{adjust}{different methods to correct the p-values for multiple testing. Default is SLIM.}

\item{effect}{methdod to calculate the mean methylation different between groups
using read coverage as weights (default). When set to "mean", the generic mean is applied
and when set to "predicted", a GLM is used instead.}

\item{parShrinkNM}{a list for squeezeVar().}

\item{test}{the statistical test used to determine the methylation differences.
The F-test is used by default, the Chisq-test is also supported.}

\item{mc.cores}{integer for denoting how many cores should be used for
                 differential methylation calculations (only can be used in
                 machines with multiple cores)}
}
\value{
a methylDiff object containing the differential methylation
                     statistics and locations
}
\description{
The function calculates differential methylation statistics between two groups
of samples. The function uses either logistic regression test
or Fisher's Exact test to calculate differential methylation.
See references for detailed explanation on statistics.
}
\section{Details}{

 The function either uses a logistic regression
 (when there are multiple samples per group) or fisher's exact
 when there is one sample per group.
}
\examples{
data(methylKit)

# Logistic regression test will be applied since there are multiple samples in each group
# in methylBase.obj object
my.diffMeth=calculateDiffMeth(methylBase.obj,covariates=NULL,overdispersion=c("none"),adjust=c("SLIM"),
                              effect=c("wmean"),parShrinkNM=list(),test=c("F"),mc.cores=1)

# pool samples in each group
pooled.methylBase=pool(methylBase.obj,sample.ids=c("test","control"))

# After applying pool() function, there is one sample in each group.
# Fisher's exact test will be applied for differential methylation
my.diffMeth2=calculateDiffMeth(pooled.methylBase,covariates=NULL,overdispersion=c("none"),adjust=c("SLIM"),
                              effect=c("wmean"),parShrinkNM=list(),test=c("F"),mc.cores=1)
}
\references{
Altuna Akalin, Matthias Kormaksson, Sheng Li,
            Francine E. Garrett-Bakelman, Maria E. Figueroa, Ari Melnick,
            Christopher E. Mason. (2012).
            "methylKit: A comprehensive R package for the analysis
            of genome-wide DNA methylation profiles." Genome Biology.
}
\seealso{
\code{\link[methylKit]{pool}}, \code{\link[methylKit]{reorganize}}
}

