% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffMeth.R
\docType{methods}
\name{calculateDiffMeth}
\alias{calculateDiffMeth}
\alias{calculateDiffMeth,methylBase-method}
\title{Calculate differential methylation statistics}
\usage{
calculateDiffMeth(.Object,covariates,overdispersion=c("none","MN","shrinkMN"),
        adjust=c("SLIM","holm","hochberg","hommel","bonferroni","BH","BY","fdr",
        "none","qvalue"), effect=c("wmean","mean","predicted"),parShrinkNM=list(),
        test=c("F","Chisq"),mc.cores=1)
}
\arguments{
\item{.Object}{a methylBase object to calculate differential methylation}

\item{covariates}{a data.frame containing covariates, which should be included in the test.}

\item{overdispersion}{If set to "none"(default), no overdispersion correction will be attempted.
If set to "MN", basic overdispersion correction will be applied.
(NOT IMPLEMENTED: If set to "shrinkMN", overdisperison correction with squeezeVar()
from the limma-package will be applied (not implemented as of yet).}

\item{adjust}{different methods to correct the p-values for multiple testing.
Default is SLIM from methylKit.}

\item{effect}{method to calculate the mean methylation different between groups
using read coverage as weights (default). When set to "mean", the generic mean is applied
and when set to "predicted", a linear model is used instead.}

\item{parShrinkNM}{a list for squeezeVar(). (NOT IMPLEMENTED)}

\item{test}{the statistical test used to determine the methylation differences.
The Chisq-test is used by default, while the F-test can be chosen
if overdispersion control ist applied.}

\item{mc.cores}{integer denoting how many cores should be used for parallel
             differential methylation calculations (can only be used in
             machines with multiple cores).}
}
\value{
a methylDiff object containing the differential methylation
                     statistics and locations
}
\description{
The function calculates differential methylation statistics between two groups
of samples. The function uses either logistic regression test
or Fisher's Exact test to calculate differential methylation.
See references for detailed explanation on statistics.
}
\section{Details}{

Covariates can be included in the analysis. The function will then try to separate the
influence of the covariates from the treatment effect via a linear model.
The function uses the Chisq-test per default and can only use the F-test
in conjunction with the MN-overdispersion correction.
}
\examples{
data(methylKit)

# The Chisq-test will be applied when no overdispersion control is applied.
my.diffMeth=calculateDiffMeth(methylBase.obj,covariates=NULL,overdispersion=c("none"),
                              adjust=c("SLIM"),effect=c("wmean"),parShrinkNM=list(),
                              test=c("Chisq"),mc.cores=1)

# pool samples in each group
pooled.methylBase=pool(methylBase.obj,sample.ids=c("test","control"))

# After applying pool() function, there is one sample in each group.
# The F-test will be applied for differential methylation.
my.diffMeth2=calculateDiffMeth(pooled.methylBase,covariates=NULL,overdispersion=c("none"),
                               adjust=c("SLIM"),effect=c("wmean"),test=c("F"))

# Covariates and overdispersion control:
# generate a methylBase object with age as a covariate
covariates=data.frame(age=c(30,80,30,80))
sim.methylBase<-dataSim(replicates=4,sites=1000,treatment=c(1,1,0,0),
                        covariates=covariates,
                        sample.ids=c("test1","test2","ctrl1","ctrl2"))

# Apply overdispersion correction and include covariates
# in differential methylation calculations.
my.diffMeth3<-calculateDiffMeth(sim.methylBase,
                                covariates=covariates,
                                overdispersion="MN",test="Chisq",mc.cores=1)
}
\references{
Altuna Akalin, Matthias Kormaksson, Sheng Li,
            Francine E. Garrett-Bakelman, Maria E. Figueroa, Ari Melnick,
            Christopher E. Mason. (2012).
            "methylKit: A comprehensive R package for the analysis
            of genome-wide DNA methylation profiles." Genome Biology.
}
\seealso{
\code{\link[methylKit]{pool}}, \code{\link[methylKit]{reorganize}}
         \code{\link[methylKit]{dataSim}}
}

