% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffMethDSS.R
\docType{methods}
\name{calculateDiffMethDSS}
\alias{calculateDiffMethDSS}
\title{calculate Differential Methylation with DSS}
\usage{
calculateDiffMethDSS(meth, adjust=c("SLIM","holm","hochberg","hommel",
                            "bonferroni","BH","BY","fdr","none","qvalue"), mc.cores=1)
}
\arguments{
\item{meth}{a methylBase object}

\item{adjust}{different methods to correct the p-values for multiple testing.
Default is "SLIM" from methylKit. For "qvalue" please see \code{\link[qvalue]{qvalue}}
and for all other methods see \code{\link[stats]{p.adjust}}.}

\item{mc.cores}{integer denoting how many cores should be used for parallel
differential methylation calculations (can only be used in
machines with multiple cores).}
}
\value{
a methylDiff object
}
\description{
This function provides an interface to the DSS package by Hao Wo at Emory University.
}
\section{Details}{

test
}
\examples{
data(methylKit)

dssDiffay <- calculateDiffMethDSS(methylBase.obj, adjust="SLIM", mc.cores=1)
}

