% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSim.R
\docType{methods}
\name{dataSim}
\alias{dataSim}
\title{get simulated data}
\usage{
dataSim(replicates,sites=20000,percentage=20,effect=20,alpha=0.4,beta=0.5,theta=5,
               covariates=NULL,experiment=NULL,sample.ids,assembly,context,treatment,
               destranded,resolution)
}
\arguments{
\item{replicates}{number of replicates}

\item{sites}{number of CpG sites per replicate}

\item{percentage}{percentage of sites which are affected by treatment}

\item{effect}{effect size of treatment}

\item{alpha}{shape1 parameter for beta distribution (used for substitution probabilites)}

\item{beta}{shape2 parameter for beta distribution (used for substitution probabilites)}

\item{theta}{dispersion parameter for beta distribution (used for substitution probabilites)}

\item{covariates}{data.frame containing covariates}

\item{experiment}{... not implemented}

\item{sample.ids}{character vector containing sample names}

\item{assembly}{assembly template (e.g. "hg18")}

\item{context}{experimanteal context of the data (e.g. "CpG")}

\item{treatment}{vector containing treatment information}

\item{destranded}{boolean parameter}

\item{resolution}{resolution of the data (e.g. "base")}
}
\value{
a methylBase object
}
\description{
get simulated data
}
\examples{
data(methylKit)
## Following
my.methylBase=dataSim(replicates=4,sample.ids=c("test1","test2","ctrl1","ctrl2"),assembly="hg18",
                       context="CpG",treatment=c(1,1,0,0),destranded=FALSE,resolution="base")
}

