% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSim.R
\docType{methods}
\name{dataSim}
\alias{dataSim}
\title{Simulate DNA methylation data}
\usage{
dataSim(replicates,sites,treatment,percentage=20,effect=50,alpha=0.4,beta=0.5,
        theta=50,covariates=NULL,sample.ids=NULL,assembly="hg18",context="CpG")
}
\arguments{
\item{replicates}{the number of samples that should be simulated.}

\item{sites}{the number of CpG sites per sample.}

\item{treatment}{a vector containing treatment information.}

\item{percentage}{the proportion of sites which should be affected by the treatment.}

\item{effect}{a number or vector specifying the effect size of the treatment.
See `Examples'.}

\item{alpha}{shape1 parameter for beta distribution (used for substitution probabilites)}

\item{beta}{shape2 parameter for beta distribution (used for substitution probabilites)}

\item{theta}{dispersion parameter for beta distribution (used for substitution probabilites)}

\item{covariates}{a data.frame containing covariates (optional)}

\item{sample.ids}{will be generated automatically from \code{treatment}, but can be 
overwritten by a character vector containing sample names.}

\item{assembly}{the assembly description (e.g. "hg18")}

\item{context}{the experimanteal context of the data (e.g. "CpG")}
}
\value{
a methylBase object containing simulated methylation data.
}
\description{
The function simulates DNA methylation data from multiple samples.
See references for detailed explanation on statistics.
}
\section{Details}{

While the coverage is modeled with a binomial distribution, the function uses 
a Beta distribution to simulate the methylation background across all samples.\cr
The parameters \code{alpha}, \code{beta} and \code{theta} determine this beta distribution and 
thereby the methylation values.\cr
The parameters \code{percentage} and \code{effect} determine the proportion of sites that are 
affected by the treatment and the strength of this influence, respectively.\cr
The additional information needed for a valid methylBase.obj is generated as "dummy
values", but can be overwritten as needed.
}
\examples{

data(methylKit)

# Simualte data for 4 samples with 20000 sites each.
# The methylation in 10\% of the sites are elevated by 50\%.
my.methylBase=dataSim(replicates=4,sites=2000,treatment=c(1,1,0,0),
percentage=10,effect=50)

# Simulate data with variable effect sizes of the treatment
# The methylation in 30\% of the sites are elevated by 40\%, 50\% or 60\%.
my.methylBase2=dataSim(replicates=4,sites=2000,treatment=c(1,1,0,0),
percentage=30,effect=c(40,50,60),
sample.ids=c("treated1","treated2","untreated1","untreated2"))

}

