% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\docType{methods}
\name{annotate.WithFeature}
\alias{annotate.WithFeature}
\alias{annotate.WithFeature.Flank}
\alias{annotate.WithGenicParts}
\alias{read.bed}
\alias{read.feature.flank}
\alias{read.transcript.features}
\title{Deprecated/Defunct functions}
\usage{
annotate.WithFeature()

annotate.WithFeature.Flank()

annotate.WithGenicParts()

read.bed()

read.feature.flank()

read.transcript.features()
}
\description{
These are deprecated or defunct functions. Most of them 
are replaced by genomation functions.
}

