% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffMeth.R
\docType{methods}
\name{get.methylDiff}
\alias{get.methylDiff}
\alias{get.methylDiff,methylDiff-method}
\title{get differentially methylated regions/bases based on cutoffs}
\usage{
get.methylDiff(.Object,difference=25,qvalue=0.01,type="all")

\S4method{get.methylDiff}{methylDiff}(.Object, difference = 25,
  qvalue = 0.01, type = "all")
}
\arguments{
\item{.Object}{a \code{\link{methylDiff}} object}

\item{difference}{cutoff for absolute value of methylation percentage change between test and control (default:25)}

\item{qvalue}{cutoff for qvalue of differential methylation statistic (default:0.01)}

\item{type}{one of the "hyper","hypo" or "all" strings. Specifies what type of differentially menthylated bases/regions should be returned.
For retrieving Hyper-methylated regions/bases type="hyper", for hypo-methylated type="hypo" (default:"all")}
}
\value{
a methylDiff object containing the differential methylated locations satisfying the criteria
}
\description{
The function subsets a \code{\link{methylDiff}} object in order to get 
differentially methylated bases/regions
satisfying thresholds.
}
\examples{

data(methylKit)

# get differentially methylated bases/regions with specific cutoffs
all.diff=get.methylDiff(methylDiff.obj,difference=25,qvalue=0.01,type="all")

# get hyper-methylated
hyper=get.methylDiff(methylDiff.obj,difference=25,qvalue=0.01,type="hyper")

# get hypo-methylated
hypo=get.methylDiff(methylDiff.obj,difference=25,qvalue=0.01,type="hypo")


}

