% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backbone.R, R/methylDBClasses.R
\docType{methods}
\name{getCorrelation}
\alias{getCorrelation}
\alias{getCorrelation,methylBase-method}
\alias{getCorrelation,methylBaseDB-method}
\title{get correlation between samples in methylBase or methylBaseDB object}
\usage{
getCorrelation(object,method="pearson",plot=FALSE,nrow)

\S4method{getCorrelation}{methylBase}(object, method, plot)

\S4method{getCorrelation}{methylBaseDB}(object, method = "pearson",
  plot = FALSE, nrow = 2e+06)
}
\arguments{
\item{object}{a methylBase or methylBaseDB object}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed (default:"pearson", other options are "kendall" and "spearman")}

\item{plot}{scatterPlot if TRUE (default:FALSE)}

\item{nrow}{a numeric giving the number of lines to read in of methylBaseDB object, defaults to 2e6}
}
\value{
a correlation matrix object and plot scatterPlot
}
\description{
The functions returns a matrix of correlation coefficients and/or a set of scatterplots showing the relationship between samples
}
\section{Details}{
 The argument 'nrow' is only evaluated if the input is a \code{methylBaseDB} object.
When 'nrow' is not specified \code{getCorrelation} will read the first 2M records of the given object,
but if you want to read all records 'nrow' has to be NULL.
You should change 'nrow' if using \code{getCorrelation} with all records of the methylBaseDB object would take to long.
}
\examples{
data(methylKit)
getCorrelation(methylBase.obj,method="pearson",plot=FALSE)
getCorrelation(methylBaseDB.obj,method="pearson",plot=FALSE,nrow=10000)
}

