\docType{methods}
\name{getCoverageStats}
\alias{"getCoverageStats,-methods"}
\alias{"getCoverageStats,methylRaw-method"}
\alias{getCoverageStats}
\title{get coverage stats from methylRaw object}
\arguments{
  \item{.Object}{a \code{methylRaw} object}

  \item{plot}{plot a histogram of coverage if TRUE
  (deafult:FALSE)}

  \item{both.strands}{do stats and plot for both strands if
  TRUE (deafult:FALSE)}
}
\value{
  a summary of coverage statistics or plot a histogram of
  coverage
}
\description{
  get coverage stats from methylRaw object
}

