\docType{methods}
\name{getData}
\alias{"getData,-methods"}
\alias{"getData,ANY-method"}
\alias{"getData,methylBase-method"}
\alias{"getData,methylRaw-method"}
\alias{getData}
\title{gets the data slot from the methylBase object}
\arguments{
  \item{x}{a methylBase object}
}
\value{
  data.frame for methylation events
}
\description{
  The data retrived from this function is of a
  \code{data.frame}. This is basically containing all
  relevant methylation information per region
}

