\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,-methods}
\alias{getData,methylBase-method}
\alias{getData,methylDiff-method}
\alias{getData,methylRaw-method}
\alias{getData-method}
\title{gets the data slot from the methylBase object}
\arguments{
  \item{x}{a methylBase object}
}
\value{
  data.frame for methylation events
}
\description{
  The data retrived from this function is of a
  \code{data.frame}. This is basically containing all
  relevant methylation information per region
}

