\docType{methods}
\name{getMembers}
\alias{getMembers}
\alias{getMembers,annotationByFeature-method}
\title{Get the membership slot of annotationByFeature}
\arguments{
  \item{x}{a \code{annotationByFeature} object}
}
\value{
  RETURNS a matrix showing overlap of target features with
  annotation features. 1 for overlap, 0 for non-overlap
}
\description{
  Membership slot defines the overlap of target features
  with annotation features For example, if a target feature
  overlaps with an exon
}

