% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylDBClasses.R
\docType{methods}
\name{makeMethylDB}
\alias{makeMethylDB}
\alias{makeMethylDB,methylBase-method}
\alias{makeMethylDB,methylBase-methods}
\alias{makeMethylDB,methylRaw-method}
\alias{makeMethylDB,methylRaw-methods}
\alias{makeMethylDB,methylDiff-method}
\alias{makeMethylDB,methylDiff-methods}
\alias{makeMethylDB,methylRawList-method}
\alias{makeMethylDB,methylRawList-methods}
\title{load tabix file with header to methylDB}
\usage{
makeMethylDB(obj, dbdir = getwd())

\S4method{makeMethylDB}{methylBase}(obj, dbdir = getwd())

\S4method{makeMethylDB}{methylRaw}(obj, dbdir = getwd())

\S4method{makeMethylDB}{methylDiff}(obj, dbdir = getwd())

\S4method{makeMethylDB}{methylRawList}(obj, dbdir = getwd())

makeMethylDB(obj, dbdir = getwd())

\S4method{makeMethylDB}{methylBase}(obj, dbdir = getwd())

\S4method{makeMethylDB}{methylRaw}(obj, dbdir = getwd())

\S4method{makeMethylDB}{methylDiff}(obj, dbdir = getwd())

\S4method{makeMethylDB}{methylRawList}(obj, dbdir = getwd())
}
\arguments{
\item{dbpath}{path to a tabix file with header}
}
\value{
an \code{\link{methylBaseDB}},\code{\link{methylRawDB}},
\code{\link{methylRawListDB}} or an \code{\link{methylDiffDB}} object
}
\description{
load tabix file with header to methylDB
}
\details{
The function reads the header from a given tabix file and 
loads it into corresponding methylDB object.
}
\examples{
\dontrun{
data(methylKit)


makeMethylDB(methylBase.obj,"my/path")
}

}
