\docType{data}
\name{methylBase.obj}
\alias{methylBase.obj}
\title{Example methylBase object.}
\format{methylBase.obj object stores the location and methylation information for bases that
are covered in all samples. \code{methylBase} partially extends \code{data.frame} S3 class.}
\description{
  methylKit has several objects. This data set includes
  examples of the following objects: \code{methylBase},
  \code{methylDiff} and \code{methylRawList}. You can load
  the data using \code{data(methylKit)}
}

