\docType{data}
\name{methylDiff.obj}
\alias{methylDiff.obj}
\title{Example methylKit objects.}
\format{The Differential methylation information is stored in methylDiff.obj object.
\code{methylBase} partially extends \code{data.frame} S3 class.}
\description{
  methylKit has several objects. This data set includes
  examples of the following objects: \code{methylBase},
  \code{methylDiff} and \code{methylRawList}. You can load
  the data using \code{data(methylKit)}
}

