% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{methylDiff.obj}
\alias{methylDiff.obj}
\title{Example methylKit objects.}
\format{The Differential methylation information is stored in 
methylDiff.obj object.
  \code{methylBase} partially extends \code{data.frame} S3 class.}
\description{
\code{methylBase}, \code{methylDiff} and \code{methylRawList}. 
 You can load the data using \code{data(methylKit)}
}
