% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{methylRawList.obj}
\alias{methylRawList.obj}
\title{Example methylRawList object.}
\format{Methylation data from multiple the samples regardless 
  of common coverage are stored in  methylRawList.obj object. \code{methylRawList} extends \code{list} S3 class}
\description{
\code{methylBase}, \code{methylDiff} and \code{methylRawList}. You can load the data using \code{data(methylKit)}
}

