\docType{methods}
\name{PCASamples}
\alias{"PCASamples,"}
\alias{"PCASamples,-methods"}
\alias{"PCASamples,ANY-method"}
\alias{"methylBase-method"}
\alias{PCASamples}
\title{Principal Components Analysis on samples in methylBase object}
\arguments{
  \item{.Object}{a \code{methylBase} object}

  \item{cor}{a logical value indicating whether the
  calculation should use the correlation matrix or the
  covariance matrix. (The correlation matrix can only be
  used if there are no constant variables.)}
}
\value{
  The form of the value returned by \code{PCASamples} is
  the table of importance of components
}
\description{
  Principal Components Analysis on samples in methylBase
  object
}

