\docType{methods}
\name{percMethylation}
\alias{percMethylation}
\alias{percMethylation,methylBase-method}
\title{get percent methylation scores from methylBase object}
\usage{
  percMethylation(methylBase.obj)
}
\arguments{
  \item{methylBase.obj}{a methylBase object}
}
\value{
  matrix with percent methylation values per base/region
  across all samples, row names would be base/region
  identifiers
}
\description{
  get percent methylation scores from methylBase object
}

