\docType{methods}
\name{select}
\alias{select}
\alias{select,methylBase-method}
\alias{select,methylDiff-method}
\alias{select,methylRaw-method}
\title{selects rows from of methylRaw.methylBase and methylDiff objects}
\description{
  selects rows from of methylRaw.methylBase and methylDiff
  objects
}
\examples{
# select(methylRaw.obj,1:100) # selects first hundred rows, returns a methylRaw object
 # select(methylBase.obj,1:100)
 # select(methylDiff.obj,1:100)
}

