% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.R, R/backbone.R, R/diffMeth.R
\docType{methods}
\name{show,annotationByGenicParts-method}
\alias{show,annotationByFeature-method}
\alias{show,annotationByGenicParts-method}
\alias{show,methylBase}
\alias{show,methylBase-method}
\alias{show,methylDiff}
\alias{show,methylDiff-method}
\alias{show,methylRaw}
\alias{show,methylRaw-method}
\alias{show,methylRawList}
\alias{show,methylRawList-method}
\title{show method for methylKit classes}
\usage{
\S4method{show}{annotationByGenicParts}(object)

\S4method{show}{annotationByFeature}(object)

\S4method{show}{methylBase}(object)

\S4method{show}{methylRaw}(object)

\S4method{show}{methylRawList}(object)

\S4method{show}{methylDiff}(object)
}
\description{
The show method works for \code{methylRaw},\code{methylRawList},
\code{methylBase} and \code{methylDiff} objects
}
\examples{
data(methylKit)
methylDiff.obj
show(methylDiff.obj)
}

