\name{show}
\alias{show}
\alias{show,annotationByFeature-method}
\alias{show,annotationByGenicParts-method}
\alias{show,methylBase}
\alias{show,methylDiff}
\alias{show,methylRaw}
\alias{show,methylRawList}
\title{show method for methylKit classes}
\description{
  The show method works for
  \code{methylRaw},\code{methylRawList}, \code{methylBase}
  and \code{methylDiff} objects
}
\examples{
data(methylKit)
methylDiff.obj
show(methylDiff.obj)
}

