% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backbone.R, R/diffMeth.R, R/methylDBClasses.R
\docType{methods}
\name{show,methylBase-method}
\alias{show,methylBase-method}
\alias{show,methylBase}
\alias{show,methylRaw-method}
\alias{show,methylRaw}
\alias{show,methylRawList-method}
\alias{show,methylRawList}
\alias{show,methylDiff-method}
\alias{show,methylDiff}
\alias{show,methylRawDB-method}
\alias{show,methylRawDB}
\alias{show,methylRawListDB-method}
\alias{show,methylRawListDB}
\alias{show,methylBaseDB-method}
\alias{show,methylBaseDB}
\alias{show,methylDiffDB-method}
\alias{show,methylDiffDB}
\title{show method for methylKit classes}
\usage{
\S4method{show}{methylBase}(object)

\S4method{show}{methylRaw}(object)

\S4method{show}{methylRawList}(object)

\S4method{show}{methylDiff}(object)

\S4method{show}{methylRawDB}(object)

\S4method{show}{methylRawListDB}(object)

\S4method{show}{methylBaseDB}(object)

\S4method{show}{methylDiffDB}(object)
}
\arguments{
\item{object}{any methylKit object}
}
\description{
The show method works for \code{methylRaw},\code{methylRawDB},
\code{methylRawList},\code{methylRawListDB},
\code{methylBase},\code{methylBaseDB} and \code{methylDiff} objects
}
\examples{
data(methylKit)
methylDiff.obj
show(methylDiff.obj)



}
