% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalize.R
\docType{methods}
\name{tileMethylCounts}
\alias{tileMethylCounts}
\alias{tileMethylCounts,methylBase-method}
\alias{tileMethylCounts,methylRaw-method}
\alias{tileMethylCounts,methylRawList-method}
\title{Get methylated/unmethylated base counts for tilling windows}
\usage{
tileMethylCounts(object,win.size=1000,step.size=1000,cov.bases=0)

\S4method{tileMethylCounts}{methylRaw}(object, win.size = 1000,
  step.size = 1000, cov.bases = 0)

\S4method{tileMethylCounts}{methylRawList}(object, win.size = 1000,
  step.size = 1000, cov.bases = 0)

\S4method{tileMethylCounts}{methylBase}(object, win.size = 1000,
  step.size = 1000, cov.bases = 0)
}
\arguments{
\item{object}{\code{\link{methylRaw}}, \code{\link{methylRawList}} or \code{\link{methylBase}} 
object containing base pair resolution methylation information}

\item{win.size}{an integer for the size of the tiling windows}

\item{step.size}{an integer for the step size of tiling windows}

\item{cov.bases}{minimum number of bases to be covered in a given window}
}
\value{
\code{methylRaw},\code{methylBase} or \code{methylRawList} object
}
\description{
The function summarizes methylated/unmethylated base counts over tilling 
windows accross genome. This function can be used when differential 
methylated analysis is preferable to tilling windows instead of base pairs.
}
\examples{
data(methylKit)

tiled.methylRaw=tileMethylCounts(object=methylRawList.obj,win.size=1000,
                                 step.size=1000,cov.bases=0)


}

