\docType{methods}
\name{updateMethObject}
\alias{updateMethObject}
\alias{updateMethObject,methylBase-method}
\alias{updateMethObject,methylDiff-method}
\alias{updateMethObject,methylRaw-method}
\alias{updateMethObject,methylRawList-method}
\title{update methylKit objects}
\usage{
  updateMethObject(object)
}
\arguments{
  \item{object}{a methylKit object: methylRaw,
  methylRawList, methylBase or methylDiff}
}
\value{
  \code{\link{methylRaw}},\code{\link{methylDiff}},\code{\link{methylBase}}
  or \code{\link{methylRawList}} object
}
\description{
  The method updates object from earlier versions (<v1.0.0)
  to latest object.
}

